requirements.txt
streamlit==1.33.0
openai==1.15.0
google-generativeai==0.3.1